Rename renderUserEvents -> renderHelpers
