+++ lib/userEvents/renderHelpers.js
