+++ lib/userEvents/reportOnUserEvents.js
@@ -1,7 +1,7 @@
 const describeTime = require('../describeTime');
 
-const renderUserEvent =  require('./renderUserEvents').renderUserEvent;
-const renderIndex = require('./renderUserEvents').renderIndex;
+const renderUserEvent =  require('./renderHelpers').renderUserEvent;
+const renderIndex = require('./renderHelpers').renderIndex;
 
 const EventTypes = {
   CreateEvent: 'CreateEvent',
